+++ src/options/Options.ts
@@ -38,6 +38,15 @@ export default class Options {
   latencyCheckInterval: number = 5000;
 
   /**
+   * The max latency, in milliseconds, a player can have before being kicked.
+   *
+   * @property {number}
+   *
+   * @default 300
+   */
+  maxLatency: number = 300;
+
+  /**
    * 
    * Defines the close codes and reasons for those codes that can be used for various actions such as closing, kicking, and banning.
    *