+++ tests/custom/server.js
@@ -23,14 +23,10 @@ const server = app.listen(3000, () => console.log('Listening on port 3000'));
 const options = {
   dbType: 'mongodb',
   pingInterval: 1000,
-  socketCloseEvents: {
-    banned: { code: 5555, reason: 'you banned fool' }
-  }
+  latencyCheckInterval: 1000
 };
 
 const gg = new GameGuard(server, options);
-
-console.log(gg._options);
 
 gg.players.connected.add(player => {